pyproject.toml
[project]
name = "patient-management-system"
version = "1.0.0"
description = "Production Ready Patient Management API"
authors = [
    {name = "Pramshu Sharma",email = "pramshu.sharma35@gmail.com"}
]
license = {text = "Public Domain"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-restx (>=1.3.2,<2.0.0)",
    "alembic (>=1.18.0,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
